Add files via upload
The nexus of anthropogenic factors and the environment have been widely studied in the past few decades since the onset of global warming, as the rise in global temperature could inflict disastrous consequences on mankind. Corroborating that the environmental changes are the product of various anthropogenic features such as population size, technological advancement, and affluence, the STIRPAT models were proposed and have been extensively employed to study the impact of socio-economic factors on the environment. This however has remained poorly unexplored for India, which has complex and varying demographic properties. Here, through the STIRPAT model, we explore the impact of anthropogenic factors -population size, percentage of the urban population, per capita GDP, and its share of service and industrial sector- on the outdoor ambient concentration of sulfur dioxide and nitrogen dioxide within the industrial and residential clusters of urban states of India. Our results indicate that growing population size has an adverse effect on the environment in every type of setup. Moreover, we show that the up-gradation in technology has a positive impact on the environment quality in industrial clusters however, there is still a need for more robust and effective policies to improve the air quality in the residential areas. Although we obtain penetrating insights into the nexus of human influence and environment quality, we also emphasize improving the quality and land coverage of ambiance data collection in order to obtain deeper insight and to effectively address the environmental crisis in India.